Refactor drive train command and add a unit test for RotateInPlace
