--- command/__init__.py
@@ -1,8 +1,0 @@
-from command.ballpath import BallPath
-from command.elevator import ElevatorDown, ElevatorSolenoidToggle, ElevatorSetupCommand, ElevatorClimbCommand, \
-    ElevatorRezero
-from command.index import IndexOn, IndexOff, IndexDrive, IndexAutoDrive
-from command.intake import IntakeAutoEject, IntakeDinglebobOn, IntakeDinglebobOff, IntakeToggleLeft, IntakeToggleRight, \
-    IntakeToggleReverse
-from command.shooter import ShooterEnable, ShooterZero, ShooterEnableAtDistance, TurretAim, NaiveDemoShot, TurretDriveAim
-from command.drivetrain import DriveSwerveTurretAim
