+++ commands/drivetrain/constants.py
@@ -1,0 +1,6 @@
+# TODO perhaps move these into a global constants file
+AIM_kP = 3.5
+AIM_kI = 0
+AIM_kD = 0
+AIM_max_angular_vel = 6
+AIM_max_angular_accel = 6
