+++ commands/drivetrain/util.py
@@ -1,0 +1,8 @@
+def curve_abs(x):
+    return x ** 2.4
+
+
+def curve(x):
+    if x < 0:
+        return -curve_abs(-x)
+    return curve_abs(x)
