+++ pyproject.toml
@@ -10,6 +10,8 @@ robotpy-toolkit-7407 = "0.5.1"
 numpy = "^1.22.2"
 opencv-python = "^4.5.5"
 serial = "^0.0.97"
+robotpy-commands-v2 = "^2022.4.1.1"
+robotpy-ctre = "^2022.1.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
